Big port to Aiko to handle each task's run period independently.
This code was a hack on top of a hack, and re-used some 1D demos that
were not meant to be run as part of a periodic callback.
As a result, they called delay from within a loop, and then I hacked up
IR reading and later 2D matrix updates within that delay.
It was a terrible hack which accumulated technical debt.

This converts the code to Aiko which calls each handler at the interval
they need to.
Aiko isn't magic though, if the 2D frame calculation takes too long to
render and/or display, it will slow down the 1D LED display as it can't
get scheduled until the 2D frame handler has returned from a single
frame calculation.

I had to rewrite all the 1D demos to remove the for loops and delays and
rely on an internal "come back to me in X milliseconds" counter.

IR handler now only does what it was supposed to: serial (and it also
reads from serial at the same time because it's convenient).

Added a warning that reading writing to PSRAM on a non PSRAM board,
will crash it hard.